criado endpoint relatorio financeiro, mas messagem 406 not acceptable, naõ consegui implementar com sucesso a consulta

+++ src/main/java/com/example/projeto_rp_condominio_dev/dto/HabitanteDTO.java
@@ -4,8 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
-
 @Getter
 @Setter
 @AllArgsConstructor
@@ -19,5 +17,5 @@ public class HabitanteDTO {
 
     private String dtNasc;
 
-    private BigDecimal renda;
+    private Double renda;
 }
